Consolidate release info constants in releaseInfoHelpers
