+++ src/core/utilities/auto-match-logic.ts
@@ -217,5 +217,3 @@ export const findMostCommonShowName = (showList: (PathDetails | null)[]) => {
     showNames[0],
   );
 };
-
-export const AUTO_MATCH_EPISODE_ID = -1;
